coap_client: split libcoap specific code
Split libcoap specific code and move it into separate module. This makes
room for Zephyr CoAP library as an alternative CoAP backend implementation.
