+++ port/linux/golioth_sdk/CMakeLists.txt
@@ -37,6 +37,7 @@ set(sdk_srcs
     "${sdk_port}/linux/fw_update_linux.c"
     "${sdk_src}/golioth_status.c"
     "${sdk_src}/coap_client.c"
+    "${sdk_src}/coap_client_libcoap.c"
     "${sdk_src}/log.c"
     "${sdk_src}/lightdb_state.c"
     "${sdk_src}/stream.c"
